ajout feuille saisie

+++ controleur/Joueur/GestionSupprimerJoueur.php
@@ -1,19 +1,17 @@
 <?php
-
-require_once __DIR__ . '/../../modele/Joueur.php';
 require_once __DIR__ . '/../../modele/dao/DaoJoueur.php';
-require_once __DIR__ . '/../../modele/dao/requetes/RequetesJoueur.php';
 require_once __DIR__ . '/../../bd/pdo.php';
-require_once __DIR__ . '/../../connexion/verificationConnexion.php'; 
+require_once __DIR__ . '/../../connexion/verificationConnexion.php';
 
 $dao = new DaoJoueur($linkpdo);
-
 $id = $_GET['id'] ?? '';
-
 $joueur = $dao->findById($id);
-if ($joueur) {
-    $dao->delete($joueur);
-}
 
-header("Location: liste.php");
-exit;
+if ($joueur && $dao->peutSupprimer($joueur)) {
+    $dao->delete($joueur);
+    header("Location: ../../vue/Joueur/liste.php?msg=ok");
+    exit;
+} else {
+    header("Location: ../../vue/Joueur/liste.php?msg=error");
+    exit;
+}
